+++ Unity/Assets/Model/Module/Message/MessageDispatcherComponent.cs
@@ -26,7 +26,7 @@ namespace ET
 	/// </summary>
 	public class MessageDispatcherComponent : Entity
 	{
-		private readonly Dictionary<ushort, List<IMHandler>> handlers = new Dictionary<ushort, List<IMHandler>>();
+		private readonly Dictionary<uint, List<IMHandler>> handlers = new Dictionary<uint, List<IMHandler>>();
 
 		public void Awake()
 		{
@@ -55,7 +55,7 @@ namespace ET
 				}
 
 				Type messageType = iMHandler.GetMessageType();
-				ushort opcode = this.Parent.GetComponent<OpcodeTypeComponent>().GetOpcode(messageType);
+				uint opcode = this.Parent.GetComponent<OpcodeTypeComponent>().GetOpcode(messageType);
 				if (opcode == 0)
 				{
 					Log.Error($"消息opcode为0: {messageType.Name}");
@@ -65,7 +65,7 @@ namespace ET
 			}
 		}
 
-		public void RegisterHandler(ushort opcode, IMHandler handler)
+		public void RegisterHandler(uint opcode, IMHandler handler)
 		{
 			if (!this.handlers.ContainsKey(opcode))
 			{
